added utilities folder, namespaced loading and saving data into FileUtils module
