+++ scripts/utils/validate.rb
@@ -1,12 +1,3 @@
-class Numeric
-  def to_digits(n = 3)
-    str = self.to_s
-    return str if str.size >= n
-    str.prepend("0" * (n - str.size))
-    return str
-  end
-end
-
 def validate(hash)
   errors = hash.map do |key, value|
     if value.is_a?(Array)
